🚜!HOTFIX: 패키지 경로 변경
1. 패키지에 domain 경로 추가

Resolves: #14
