+++ src/main/kotlin/com/sparta/reviewus/domain/exception/DuplicatedPropertyException.kt
@@ -1,0 +1,4 @@
+package com.sparta.reviewus.domain.exception
+
+data class DuplicatedPropertyException(val property: String, val value: String) :
+    RuntimeException("해당 $property $value 은 이미 존재하는 이메일 입니다.")
