+++ src/main/kotlin/com/sparta/reviewus/domain/exception/ModelNotFoundException.kt
@@ -1,0 +1,4 @@
+package com.sparta.reviewus.domain.exception
+
+data class ModelNotFoundException(val modelName: String, val id: Long?) :
+    RuntimeException("Model $modelName not found with given id: $id")
