+++ src/main/kotlin/com/sparta/reviewus/domain/exception/dto/ErrorResponse.kt
@@ -1,0 +1,5 @@
+package com.sparta.reviewus.domain.exception.dto
+
+data class ErrorResponse(
+    val message: String?
+)
