+++ src/main/kotlin/com/sparta/reviewus/domain/feed/dto/CreateFeedResponse.kt
@@ -1,0 +1,15 @@
+package com.sparta.reviewus.domain.feed.dto
+
+import com.sparta.reviewus.domain.member.dto.MemberResponse
+
+data class CreateFeedResponse(
+    val id : Long,
+    val member : MemberResponse,
+    val category : String,
+    val title : String,
+    val feedPicUrl : String,
+    val longitude : String?,
+    val latitude : String?,
+    val description : String?
+)
+
