+++ src/main/kotlin/com/sparta/reviewus/domain/feed/repository/LikeRepository.kt
@@ -1,0 +1,11 @@
+package com.sparta.reviewus.domain.feed.repository
+
+import com.sparta.reviewus.domain.feed.model.Like
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface LikeRepository : JpaRepository<Like, Long>{
+
+    fun findByMemberIdAndFeedId(memberId: Long, feedId: Long): Like?
+    fun existsByMemberIdAndFeedId(memberId: Long, feedId: Long): Boolean
+    fun deleteByMemberIdAndFeedId(memberId: Long, feedId: Long): Int
+}
