+++ src/main/kotlin/com/sparta/reviewus/domain/feed/service/LikeService.kt
@@ -1,0 +1,7 @@
+package com.sparta.reviewus.domain.feed.service
+
+interface LikeService {
+
+    fun like(memberId: Long, feedId: Long): Boolean
+
+}
