+++ src/main/kotlin/com/sparta/reviewus/domain/member/service/MemberService.kt
@@ -1,0 +1,17 @@
+package com.sparta.reviewus.domain.member.service
+
+import com.sparta.reviewus.domain.member.dto.JoinRequest
+import com.sparta.reviewus.domain.member.dto.MemberResponse
+import com.sparta.reviewus.domain.member.dto.ProfileUpdateRequest
+
+interface MemberService {
+
+    fun getMemberProfile(): MemberResponse
+
+    fun getOtherMember(memberId: Long): MemberResponse
+
+    fun updateMemberProfile(profileUpdateRequest: ProfileUpdateRequest): MemberResponse
+
+    fun join(joinRequest: JoinRequest): String
+
+}
