--- src/main/kotlin/com/sparta/reviewus/exception/DuplicatedPropertyException.kt
@@ -1,4 +1,0 @@
-package com.sparta.reviewus.exception
-
-data class DuplicatedPropertyException(val property: String, val value: String) :
-    RuntimeException("해당 $property $value 은 이미 존재하는 이메일 입니다.")
