--- src/main/kotlin/com/sparta/reviewus/exception/GlobalExceptionHandler.kt
@@ -1,41 +1,0 @@
-package com.sparta.reviewus.exception
-
-import com.sparta.reviewus.exception.dto.ErrorResponse
-import com.sparta.reviewus.exception.member.BadCredentialsException
-import org.springframework.http.HttpStatus
-import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.MethodArgumentNotValidException
-import org.springframework.web.bind.annotation.ExceptionHandler
-import org.springframework.web.bind.annotation.RestControllerAdvice
-
-@RestControllerAdvice
-class GlobalExceptionHandler {
-
-    @ExceptionHandler(DuplicatedPropertyException::class)
-    fun handleDuplicatePropertyException(e: DuplicatedPropertyException): ResponseEntity<ErrorResponse> {
-        return ResponseEntity.status(HttpStatus.CONFLICT)
-            .body(ErrorResponse(e.message))
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException::class)
-    fun handleMethodArgumentNotValidException(e: MethodArgumentNotValidException): ResponseEntity<ErrorResponse> {
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-            .body(ErrorResponse(e.bindingResult.allErrors.map { it.defaultMessage }
-                .reduce { acc, string -> "$acc $string" }))
-    }
-
-    @ExceptionHandler(ModelNotFoundException::class)
-    fun handleModelNotFoundException(e: ModelNotFoundException): ResponseEntity<ErrorResponse> {
-        return ResponseEntity
-            .status(HttpStatus.NOT_FOUND)
-            .body(ErrorResponse(e.message))
-    }
-
-    @ExceptionHandler(BadCredentialsException::class)
-    fun handleModelNotFoundException(e: BadCredentialsException): ResponseEntity<ErrorResponse> {
-        return ResponseEntity
-            .status(HttpStatus.BAD_REQUEST)
-            .body(ErrorResponse(e.message))
-    }
-
-}
