--- src/main/kotlin/com/sparta/reviewus/exception/ModelNotFoundException.kt
@@ -1,4 +1,0 @@
-package com.sparta.reviewus.exception
-
-data class ModelNotFoundException(val modelName: String, val id: Long?) :
-    RuntimeException("Model $modelName not found with given id: $id")
