--- src/main/kotlin/com/sparta/reviewus/exception/dto/ErrorResponse.kt
@@ -1,5 +1,0 @@
-package com.sparta.reviewus.exception.dto
-
-data class ErrorResponse(
-    val message: String?
-)
