--- src/main/kotlin/com/sparta/reviewus/feed/dto/CreateFeedRequest.kt
@@ -1,11 +1,0 @@
-package com.sparta.reviewus.feed.dto
-
-data class CreateFeedRequest(
-    val memberId : Long,
-    val category : String,
-    val title : String,
-    val feedPicUrl : String,
-    val longitude : String?,
-    val latitude : String?,
-    val description : String?
-)
