--- src/main/kotlin/com/sparta/reviewus/feed/model/Like.kt
@@ -1,22 +1,0 @@
-package com.sparta.reviewus.feed.model
-
-import com.sparta.reviewus.member.model.Member
-import jakarta.persistence.*
-
-@Entity
-@Table(name = "likes")
-data class Like(
-
-    @ManyToOne
-    @JoinColumn(name = "member_id", nullable = false)
-    var member: Member,
-
-    @ManyToOne
-    @JoinColumn(name = "feed_id", nullable = false)
-    var feed: Feed
-
-){
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var id: Long? = null
-}
