--- src/main/kotlin/com/sparta/reviewus/feed/repository/FeedRepository.kt
@@ -1,8 +1,0 @@
-package com.sparta.reviewus.feed.repository
-
-import com.sparta.reviewus.feed.model.Feed
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface FeedRepository : JpaRepository<Feed, Long> {
-
-}
