--- src/main/kotlin/com/sparta/reviewus/feed/service/LikeService.kt
@@ -1,7 +1,0 @@
-package com.sparta.reviewus.feed.service
-
-interface LikeService {
-
-    fun like(memberId: Long, feedId: Long): Boolean
-
-}
