--- src/main/kotlin/com/sparta/reviewus/member/dto/MemberResponse.kt
@@ -1,12 +1,0 @@
-package com.sparta.reviewus.member.dto
-
-data class MemberResponse(
-    val id: Long,
-    val email: String,
-    val name: String,
-    val profilePicUrl: String,
-    val nickname: String,
-    val introduction: String,
-    val address: String?,
-    val interest: String?,
-)
