--- src/main/kotlin/com/sparta/reviewus/member/model/Member.kt
@@ -1,40 +1,0 @@
-package com.sparta.reviewus.member.model
-
-import com.sparta.reviewus.member.dto.MemberResponse
-import jakarta.persistence.*
-
-@Entity
-@Table(name="member")
-class Member(
-
-    @Column(name="email", nullable = false)
-    val email: String,
-
-    @Column(name="name", nullable = false)
-    var name: String,
-
-    @Column(name="password", nullable = false)
-    var password: String,
-
-    @Embedded
-    var profile: Profile
-
-){
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var id: Long? = null
-
-}
-
-fun Member.toResponse(): MemberResponse {
-    return MemberResponse(
-        id = id!!,
-        email = email,
-        name = name,
-        nickname = profile.nickname,
-        profilePicUrl = profile.profilePicUrl,
-        introduction = profile.introduction,
-        address = profile.address,
-        interest = profile.interest
-    )
-}
