--- src/main/kotlin/com/sparta/reviewus/member/service/MemberService.kt
@@ -1,17 +1,0 @@
-package com.sparta.reviewus.member.service
-
-import com.sparta.reviewus.member.dto.JoinRequest
-import com.sparta.reviewus.member.dto.MemberResponse
-import com.sparta.reviewus.member.dto.ProfileUpdateRequest
-
-interface MemberService {
-
-    fun getMemberProfile(): MemberResponse
-
-    fun getOtherMember(memberId: Long): MemberResponse
-
-    fun updateMemberProfile(profileUpdateRequest: ProfileUpdateRequest): MemberResponse
-
-    fun join(joinRequest: JoinRequest): String
-
-}
